# packages/schemas/openapi.yaml
openapi: 3.1.0
info:
  title: Broker MVP API
  version: 0.1.0
  description:
    Minimal OpenAPI for the broker MVP. Data-plane endpoints require mutual TLS and a short-lived session token.
servers:
  - url: https://broker.example
security: []
components:
  securitySchemes:
    mutualTLS:
      type: mutualTLS
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
...
    Error:
      type: object
      required: [error, message, correlation_id]
      properties:
        error:
          type: string
        message:
          type: string
        correlation_id:
          type: string
      additionalProperties: false
...
    IntegrationCreateResponse:
      type: object
      required: [integration_id]
      properties:
        integration_id:
          type: string
      additionalProperties: false
    IntegrationListResponse:
      type: object
      required: [integrations]
      properties:
        integrations:
          type: array
          items:
            $ref: '#/components/schemas/Integration'
      additionalProperties: false
    IntegrationUpdateRequest:
      type: object
      properties:
        enabled:
          type: boolean
        template_id:
          type: string
      additionalProperties: false
    TemplateCreateResponse:
      type: object
      required: [template_id, version]
      properties:
        template_id:
          type: string
        version:
          type: integer
      additionalProperties: false
    TemplateListResponse:
      type: object
      required: [templates]
      properties:
        templates:
          type: array
          items:
            $ref: '#/components/schemas/Template'
      additionalProperties: false
...
    Integration:
      $ref: ./integration.schema.json
...
    Template:
      $ref: ./template.schema.json
...
paths:
...
  /v1/tenants/{tenantId}/integrations:
    post:
      summary: Create integration
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './integration-write.schema.json'
      responses:
        '201':
          description: Integration created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationCreateResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List integrations
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Integrations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/integrations/{integrationId}:
    patch:
      summary: Update integration
      security:
        - bearerAuth: []
      parameters:
        - name: integrationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntegrationUpdateRequest'
      responses:
        '200':
          description: Integration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Integration'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete integration (disable access)
      security:
        - bearerAuth: []
      parameters:
        - name: integrationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Integration deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Integration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/templates:
    post:
      summary: Upload template
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './template.schema.json'
      responses:
        '201':
          description: Template uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateCreateResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List templates
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Templates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/templates/{templateId}:
    delete:
      summary: Delete template (disable all active versions)
      security:
        - bearerAuth: []
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Template deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Template is in use and cannot be deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/templates/{templateId}/versions/{version}:
    get:
      summary: Get template version
      security:
        - bearerAuth: []
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...